refactor: reorganize src into app, shared, and feature-based modules
