--- src/pages/Home/Home.module.scss
@@ -1,17 +1,0 @@
-.layout {
-  min-height: 100vh;
-}
-
-.header {
-  background-color: #001529;
-  color: white;
-  text-align: center;
-  font-size: 1.5rem;
-}
-
-.content {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  flex-direction: column;
-}
