--- src/pages/Login/Login.module.scss
@@ -1,34 +1,0 @@
-.layout {
-  min-height: 100vh;
-}
-
-.header {
-  background-color: #001529;
-  color: white;
-  text-align: center;
-  font-size: 1.5rem;
-}
-
-.content {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
-
-.form {
-  width: 100%;
-  max-width: 400px;
-  padding: 2rem;
-  background-color: #fff;
-  border: 1px solid #dcdcdc;
-  border-radius: 8px;
-  box-shadow: 0 4px 6px rgba(0, 0, 0, 0.1);
-}
-
-.input {
-  width: 100%;
-}
-
-.loginButton {
-  width: 100%;
-}
