+++ src/shared/store/store.ts
@@ -1,6 +1,6 @@
 import { configureStore } from "@reduxjs/toolkit";
 import { apiSlice } from "../api/apiSlice";
-import authReducer from "../features/auth/authSlice";
+import authReducer from "../../features/auth/model/authSlice";
 
 export const store = configureStore({
   reducer: {
